uniclub/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-extensions==2.2.5
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-rosetta==0.9.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.5.4
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
ipdb==0.12.3
ipython==7.13.0
ipython-genutils==0.2.0
isodate==0.6.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
packaging==20.3
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.11.1
pickleshare==0.7.5
Pillow==7.0.0
polib==1.1.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pyotp==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.18
wcwidth==0.1.9
zeep==3.4.0
celery[redis]==4.4.0
django-celery-beat==2.0.0